# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "One API 是一个统一的 AI 接口管理平台,支持多种 AI 服务提供商,包括 OpenAI、Anthropic、Google、Kling 等",
        "title": "One API",
        "termsOfService": "https://github.com/songquanpeng/one-api",
        "contact": {
            "name": "API Support",
            "url": "https://github.com/songquanpeng/one-api/issues"
        },
        "license": {
            "name": "MIT",
            "url": "https://github.com/songquanpeng/one-api/blob/main/LICENSE"
        },
        "version": "1.0"
    },
    "host": "localhost:3000",
    "basePath": "/",
    "paths": {
        "/kling/internal/callback": {
            "post": {
                "description": "接收 Kling 平台的异步回调通知，自动更新视频生成任务状态和结果\n\n**注意事项**:\n- 此接口由 Kling 平台自动调用，无需手动请求\n- 不需要 Authorization 认证\n- 回调成功后会自动完成计费\n- 需要在系统配置中设置 ServerAddress 为公网可访问地址\n\n**回调时机**: 任务完成（成功或失败）时触发",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Kling视频生成"
                ],
                "summary": "Kling 回调处理（内部接口）",
                "parameters": [
                    {
                        "description": "Kling 回调数据（由 Kling 平台发送）",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "回调处理成功",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "服务器错误",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/kling/v1/videos/advanced-lip-sync": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "为视频中的人脸添加对口型效果，使人物口型与音频同步\n\n**前置要求**: 必须先调用人脸识别接口获取 session_id 和 face_id\n\n**参数说明**:\n- model: 模型名称（必填）如 \"kling-v1\"\n- session_id: 人脸识别返回的会话 ID（必填）\n- face_choose: 人脸选择数组（必填）\n\n**face_choose 参数**:\n- face_id: 人脸 ID（必填，从人脸识别接口获取）\n- audio_id 或 sound_file: 音频 ID 或音频文件 URL（二选一必填）\n- sound_start_time: 音频开始时间，毫秒（必填）\n- sound_end_time: 音频结束时间，毫秒（必填）\n- sound_insert_time: 音频插入到视频的时间点，毫秒（必填）\n- sound_volume: 音频音量 0-1（可选，默认 1.0）\n- original_audio_volume: 原始音频音量 0-1（可选，默认 1.0）\n\n**计费说明**: 任务成功后通过回调结束时根据视频时长计费（后扣费）\n\n**使用流程**: 1. 人脸识别 → 2. 创建对口型任务 → 3. 查询任务结果或等待回调",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Kling视频生成"
                ],
                "summary": "Kling 高级对口型（异步接口）",
                "parameters": [
                    {
                        "type": "string",
                        "default": "Bearer sk-xxxxx",
                        "description": "Bearer Token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "对口型请求参数",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "face_choose": {
                                    "type": "array",
                                    "items": {
                                        "type": "object",
                                        "properties": {
                                            "audio_id": {
                                                "type": "string"
                                            },
                                            "face_id": {
                                                "type": "string"
                                            },
                                            "original_audio_volume": {
                                                "type": "number"
                                            },
                                            "sound_end_time": {
                                                "type": "integer"
                                            },
                                            "sound_file": {
                                                "type": "string"
                                            },
                                            "sound_insert_time": {
                                                "type": "integer"
                                            },
                                            "sound_start_time": {
                                                "type": "integer"
                                            },
                                            "sound_volume": {
                                                "type": "number"
                                            }
                                        }
                                    }
                                },
                                "model": {
                                    "type": "string"
                                },
                                "session_id": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "对口型任务已创建，返回 task_id",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "code": {
                                    "type": "integer"
                                },
                                "data": {
                                    "type": "object",
                                    "properties": {
                                        "created_at": {
                                            "type": "integer"
                                        },
                                        "task_id": {
                                            "type": "string"
                                        },
                                        "task_status": {
                                            "type": "string"
                                        },
                                        "updated_at": {
                                            "type": "integer"
                                        }
                                    }
                                },
                                "message": {
                                    "type": "string"
                                },
                                "request_id": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "object",
                                    "properties": {
                                        "code": {
                                            "type": "string"
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "type": {
                                            "type": "string"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "认证失败",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "type": {
                                            "type": "string"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "402": {
                        "description": "余额不足",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "type": {
                                            "type": "string"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "服务器错误",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "type": {
                                            "type": "string"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "securityDefinitions": {
        "Bearer": {
            "description": "在请求头中添加 Bearer Token: Authorization: Bearer {token}",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}